Refactoring for 1.0.0 complete with new build scripts and azure pipelines.
