+++ driver-3/src/test/java/com/azure/cosmos/cassandra/package-info.java
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * Verifies the Azure Cosmos Cassandra API extensions for use with DataStax Java Driver 4.
+ */
+package com.azure.cosmos.cassandra;
