+++ examples/java-driver-app/azure-cosmos-style.xml
