+++ examples/java-driver-app/src/test/resources/log4j2-test.properties
