--- examples/src/test/java/com/microsoft/azure/cosmos/cassandra/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Shows how to use the Azure Cosmos Cassandra API extensions for DataStax Java Driver 3.
- */
-package com.microsoft.azure.cosmos.cassandra;
