--- examples/testng.xml
@@ -1,8 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="CI" verbose="1">
-  <test name="examples">
-    <classes>
-      <class name="com.microsoft.azure.cosmos.cassandra.CosmosCassandraExtensionsExample"/>
-    </classes>
-  </test>
-</suite>
