--- package/src/main/java/com/microsoft/azure/cosmos/cassandra/package-info.java
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Implements DataStax Java Driver 3 extensions that encapsulate best practices for communicating with Azure Cosmos
- * Cassandra API instances.
- */
-package com.microsoft.azure.cosmos.cassandra;
