--- package/src/test/java/com/microsoft/azure/cosmos/cassandra/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Verifies the Azure Cosmos Cassandra API extensions for use with DataStax Java Driver 4.
- */
-package com.microsoft.azure.cosmos.cassandra;
